restructured atoms folder
