+++ src/frontend/atoms/error/ErrorAtom.tsx
