+++ src/frontend/atoms/error/showErrorDialog.tsx
