+++ src/frontend/atoms/movements/selectedMovementDetails.tsx
@@ -1,5 +1,5 @@
 import { atom } from 'recoil'
-import { MovementDetailData } from '../@types/frontend.types'
+import { MovementDetailData } from '../../@types/frontend.types'
 export const defaultMovementDetails :Array<MovementDetailData> = []
 
 export const selectedMovementDetailsState = atom({
