+++ src/frontend/atoms/stock/isStockSelectedAtom.tsx
