+++ src/frontend/atoms/stock/selectedStockAtom.tsx
