+++ src/frontend/components/stock/StockTableHeader.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Button } from 'primereact/button'
 import { InputText } from 'primereact/inputtext'
-import { isStockCheckedState } from '../../atoms/isStockSelectedAtom'
+import { isStockCheckedState } from '../../atoms/stock/isStockSelectedAtom'
 import { globalFilterValueState } from '../../atoms/globalFilterValueAtom'
 import { StockTableHeaderProps } from '../../@types/frontend.types'
-import { showErrorDialogState } from '../../atoms/showErrorDialog'
+import { showErrorDialogState } from '../../atoms/error/showErrorDialog'
 import { showUpdateDialogState } from '../../atoms/showUpdateDialogAtom'
 import useDeleteStockMutation from '../../hooks/stock/useDeleteStockMutation'
 import { useRecoilState } from 'recoil'
