+++ src/frontend/components/store/StoreTableHeader.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Button } from 'primereact/button'
 import { InputText } from 'primereact/inputtext'
-import { isStoreCheckedState } from '../../atoms/isStoreCheckedAtom'
+import { isStoreCheckedState } from '../../atoms/stores/isStoreCheckedAtom'
 import { globalFilterValueState } from '../../atoms/globalFilterValueAtom'
-import { showErrorDialogState } from '../../atoms/showErrorDialog'
+import { showErrorDialogState } from '../../atoms/error/showErrorDialog'
 import { showUpdateDialogState } from '../../atoms/showUpdateDialogAtom'
 import { TableHeaderProps } from '../../@types/frontend.types'
 import useDeleteStoreMutation from '../../hooks/stores/useDeleteStoreMutation'
