+++ src/frontend/hooks/movements/useDeleteMovementMutation.tsx
@@ -1,9 +1,9 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useRecoilState } from 'recoil'
-import { defaultErrorState, ErrorState } from '../../atoms/ErrorAtom'
-import { defaultMovement, selectedMovementState } from '../../atoms/selectedMovementAtom'
-import { defaultMovementChecked, isSMovementCheckedState } from '../../atoms/setSelectedMovementAtom'
-import { showErrorDialogState } from '../../atoms/showErrorDialog'
+import { defaultErrorState, ErrorState } from '../../atoms/error/ErrorAtom'
+import { defaultMovement, selectedMovementState } from '../../atoms/movements/selectedMovementAtom'
+import { defaultMovementChecked, isSMovementCheckedState } from '../../atoms/movements/setSelectedMovementAtom'
+import { showErrorDialogState } from '../../atoms/error/showErrorDialog'
 import { deleteMovement } from '../../services/movements/deleteMovement'
 
 const useDeleteMovementMutation = (queryId: string) => {
