+++ src/frontend/hooks/stock/useDialogNewStockMutation.tsx
@@ -2,8 +2,8 @@ import { Product, Store } from '@prisma/client'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useState } from 'react'
 import { useRecoilState } from 'recoil'
-import { defaultErrorState, ErrorState } from '../../atoms/ErrorAtom'
-import { showErrorDialogState } from '../../atoms/showErrorDialog'
+import { defaultErrorState, ErrorState } from '../../atoms/error/ErrorAtom'
+import { showErrorDialogState } from '../../atoms/error/showErrorDialog'
 import { createNewStock } from '../../services/stock/createNewStock'
 import useProductsQuery from '../products/useProductsQuery'
 import useStoresQuery from '../stores/useStoresQuery'
