+++ src/frontend/hooks/stock/useDialogUpdateStockMutation.tsx
@@ -3,14 +3,14 @@ import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useEffect, useState } from 'react'
 import { useRecoilState } from 'recoil'
 import { StockUpdateData } from '../../@types/frontend.types'
-import { defaultErrorState, ErrorState } from '../../atoms/ErrorAtom'
+import { defaultErrorState, ErrorState } from '../../atoms/error/ErrorAtom'
 import {
   defaultStockChecked,
   isStockCheckedState
-} from '../../atoms/isStockSelectedAtom'
-import { defaultStock, selectedStockState } from '../../atoms/selectedStockAtom'
-import { defaultStore } from '../../atoms/selectedStoreAtom'
-import { showErrorDialogState } from '../../atoms/showErrorDialog'
+} from '../../atoms/stock/isStockSelectedAtom'
+import { defaultStock, selectedStockState } from '../../atoms/stock/selectedStockAtom'
+import { defaultStore } from '../../atoms/stores/selectedStoreAtom'
+import { showErrorDialogState } from '../../atoms/error/showErrorDialog'
 import { showUpdateDialogState } from '../../atoms/showUpdateDialogAtom'
 import { updateStock } from '../../services/stock/updateStock'
 import { findStore } from '../../services/stores/findStore'
