+++ src/frontend/hooks/stores/useDialogUpdateStoreMutation.tsx
@@ -2,13 +2,13 @@ import { Store } from '@prisma/client'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useEffect } from 'react'
 import { useRecoilState } from 'recoil'
-import { defaultErrorState, ErrorState } from '../../atoms/ErrorAtom'
+import { defaultErrorState, ErrorState } from '../../atoms/error/ErrorAtom'
 import {
   defaultStoreChecked,
   isStoreCheckedState
-} from '../../atoms/isStoreCheckedAtom'
-import { defaultStore, selectedStoreState } from '../../atoms/selectedStoreAtom'
-import { showErrorDialogState } from '../../atoms/showErrorDialog'
+} from '../../atoms/stores/isStoreCheckedAtom'
+import { defaultStore, selectedStoreState } from '../../atoms/stores/selectedStoreAtom'
+import { showErrorDialogState } from '../../atoms/error/showErrorDialog'
 import { showUpdateDialogState } from '../../atoms/showUpdateDialogAtom'
 import { updateStore } from '../../services/stores/updateStore'
 import useField from '../useField'
